requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
psycopg2==2.8.3
Flask-SQLAlchemy==2.4.1